Separated message send and message receive common methods.
