+++ e2e/test/MessageReceiveE2EMultiplexingTests.cs
@@ -180,9 +180,9 @@ namespace Microsoft.Azure.Devices.E2ETests
 
                     string payload, messageId, p1Value;
 
-                    Message msg = MessageHelper.ComposeC2DTestMessage(out payload, out messageId, out p1Value);
+                    Message msg = MessageReceive.ComposeC2DTestMessage(out payload, out messageId, out p1Value);
                     await serviceClient.SendAsync(testDevice.Id, msg).ConfigureAwait(false);
-                    await MessageHelper.VerifyReceivedC2DMessageAsync(transport, deviceClient, payload, p1Value).ConfigureAwait(false);
+                    await MessageReceive.VerifyReceivedC2DMessageAsync(transport, deviceClient, payload, p1Value).ConfigureAwait(false);
                 }
             }
             finally
