+++ e2e/test/MessageSendE2ETests.cs
@@ -199,7 +199,7 @@ namespace Microsoft.Azure.Devices.E2ETests
             using (DeviceClient deviceClient = testDevice.CreateDeviceClient(transport))
             {
                 await deviceClient.OpenAsync().ConfigureAwait(false);
-                await SendSingleMessageAndVerify(deviceClient, testDevice.Id).ConfigureAwait(false);
+                await MessageSend.SendSingleMessageAndVerifyAsync(deviceClient, testDevice.Id).ConfigureAwait(false);
                 await deviceClient.CloseAsync().ConfigureAwait(false);
             }
         }
@@ -211,28 +211,8 @@ namespace Microsoft.Azure.Devices.E2ETests
             using (DeviceClient deviceClient = testDevice.CreateDeviceClient(transportSettings))
             {
                 await deviceClient.OpenAsync().ConfigureAwait(false);
-                await SendSingleMessageAndVerify(deviceClient, testDevice.Id).ConfigureAwait(false);
+                await MessageSend.SendSingleMessageAndVerifyAsync(deviceClient, testDevice.Id).ConfigureAwait(false);
                 await deviceClient.CloseAsync().ConfigureAwait(false);
-            }
-        }
-
-        public static async Task SendSingleMessageAndVerify(DeviceClient deviceClient, string deviceId)
-        {
-            EventHubTestListener testListener = await EventHubTestListener.CreateListener(deviceId).ConfigureAwait(false);
-
-            try
-            {
-                string payload;
-                string p1Value;
-                Client.Message testMessage = MessageHelper.ComposeD2CTestMessage(out payload, out p1Value);
-                await deviceClient.SendEventAsync(testMessage).ConfigureAwait(false);
-
-                bool isReceived = await testListener.WaitForMessage(deviceId, payload, p1Value).ConfigureAwait(false);
-                Assert.IsTrue(isReceived, "Message is not received.");
-            }
-            finally
-            {
-                await testListener.CloseAsync().ConfigureAwait(false);
             }
         }
 
@@ -245,7 +225,7 @@ namespace Microsoft.Azure.Devices.E2ETests
 
                 string payload;
                 string p1Value;
-                Client.Message testMessage = MessageHelper.ComposeD2CTestMessage(out payload, out p1Value);
+                Client.Message testMessage = MessageSend.ComposeD2CTestMessage(out payload, out p1Value);
                 await moduleClient.SendEventAsync(testMessage).ConfigureAwait(false);
                 await moduleClient.CloseAsync().ConfigureAwait(false);
             }
